backend/pyproject.toml
[tool.poetry]
name = "foodgram-project-react"
version = "0.1.0"
description = "prakrikum final project"
authors = ["Nikolay Kiryanov <nikolay@kiryanov.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
django-colorfield = "^0.4.2"
djoser = "^2.1.0"
django-environ = "^0.4.5"
django-cors-headers = "^3.7.0"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.0"
pre-commit = "^2.13.0"
isort = "^5.9.2"
flake8 = "^3.9.2"
black = "^21.7b0"
django-extensions = "^3.1.3"
django-debug-toolbar = "^3.2.1"
ipython = "^7.26.0"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | config/settings
  )/
)
'''
